len argument in client callback shoudl 16 bits long instead of 8. 8 bits only allow for packets of a length of up to 255 bytes, which is less than the protocol specifies

+++ inc/smol-noice-internal.h
@@ -20,7 +20,7 @@ typedef struct smolNoice
     uint8_t clientCertLen;
     uint8_t clientPrivateKey[32];
     sc_err_t (*certCallback)(uint8_t*,uint8_t,smolcert_t*);
-    sc_err_t (*transportCallback)(uint8_t*,uint8_t);
+    sc_err_t (*transportCallback)(uint8_t*,uint16_t);
     handshakeStep handShakeStep;
     NoiseHandshakeState *handshakeState;
     NoiseCipherState *rxCipher;
